@@ -1,16 +1,18 @@
 import { CssBaseline } from '@mui/material';
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import App from './App';
 import { ThemeProvider, LanguageProvider, AuthProvider } from './contexts';
 import reportWebVitals from './reportWebVitals';
 import { SnackbarProvider } from 'notistack';
 import "./index.css";
 
-ReactDOM.render(
+const container = document.getElementById('root')
+const root = createRoot(container);
+root.render(
   <React.StrictMode>
     <CssBaseline />
-    <LanguageProvider>
+    <LanguageProvider> 
       <SnackbarProvider maxSnack={3} anchorOrigin={{ horizontal: "right", vertical: "top" }}>
         <AuthProvider>
           <ThemeProvider>
@@ -19,8 +21,7 @@ ReactDOM.render(
         </AuthProvider>
       </SnackbarProvider>
     </LanguageProvider>
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
