@@ -1,10 +1,13 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import { useEffect, useState } from 'react';
-import { db } from "../firebase";
-import { ref, getStorage, getBlob } from 'firebase/storage';
+import { db, storage } from "../firebase";
+import { ref, getBlob, listAll } from 'firebase/storage';
 import { doc, getDoc } from "firebase/firestore";
 
-const storage = getStorage();
+const getTextPromise = async ref => {
+    const blob = await getBlob(ref);
+    return await blob.text();
+};
 
 const useArticleId = (id, navigate) => {
     const [article, setArticle] = useState();
@@ -17,12 +20,11 @@ const useArticleId = (id, navigate) => {
             } else {
                 let _article = _doc.data();
                 document.title = _article.title;
-                const blob = await getBlob(ref(storage, `/articles_page/${_article.content}`));
-                _article.content = await blob.text();
+                const res = await listAll(ref(storage, `/articles_page/${_article.content}`));
+                _article.pages = await Promise.all(res.items.map(getTextPromise))
                 setArticle(_article);
             }
         };
-        
         checkRoute();
     }, []);
     
